requirements.txt
# Core ML & NLP
sentence-transformers==2.2.2
transformers==4.41.1
faiss-cpu==1.7.4
numpy==1.24.4

# Web Server
flask==3.0.2

# YAML Config
pyyaml==6.0.1

# Logging/Debugging (optional but useful)
loguru==0.7.2

# For unit testing (standard lib, but included for completeness)
unittest2==1.1.0

#dashboard
streamlit
